updated lux and temperature logic

+++ Inc/application/electrical_analyzer.h
@@ -1,15 +1,14 @@
 #pragma once
 
-
 #include <stdint.h>
 
 void init_electrical_analyzer(void);
 
 void adc_conversion_complete_callback(void);
 
-int16_t get_lux(void);
+float get_lux(void);
 
-int16_t get_temperature(void);
+float get_temperature(void);
 
 int16_t get_voltage(void);
